refactor: 기물 움직임 책임 Board -> Player
- 두 Player 가 같은 Board 를 가짐
- Board 게터 방어적 복사
- Players 삭제 및 책임 ChessGame 으로 전가
- 테스트 책임에 맞게 이동
