+++ src/test/java/chess/domain/ChessGameTest.java
@@ -7,7 +7,6 @@ import chess.domain.piece.Piece;
 import chess.domain.point.File;
 import chess.domain.point.Point;
 import chess.domain.point.Rank;
-import java.util.Map;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -16,12 +15,12 @@ class ChessGameTest {
     @Test
     @DisplayName("현재 턴의 플레이어가 기물을 움직인다.")
     void move() {
-        Map<Point, Piece> board = BoardFactory.createInitialChessBoard();
+        Board board = new Board(BoardFactory.createInitialChessBoard());
         ChessGame chessGame = new ChessGame(board);
 
         Point departure = Point.of(File.B, Rank.SECOND);
         Point destination = Point.of(File.B, Rank.THIRD);
-        chessGame.move(departure, destination);
+        chessGame.currentTurnPlayerMove(departure, destination);
 
         assertThat(board.get(departure)).isEqualTo(Piece.empty());
         assertThat(board.get(destination)).isInstanceOf(Pawn.class);
